@@ -1,12 +1,12 @@
 {
   "name": "alfio-pi",
   "version": "0.0.0",
-  "license": "MIT",
+  "license": "GPLv3",
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
     "start": "ng serve --proxy-config proxy.config.json",
-    "build": "ng build --prod --aot",
+    "build": "ng build --configuration=production --aot",
     "lint": "tslint \"src/**/*.ts\"",
     "test": "ng test",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
@@ -14,50 +14,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "4.3.2",
-    "@angular/compiler": "4.3.2",
-    "@angular/core": "4.3.2",
-    "@angular/forms": "4.3.2",
-    "@angular/http": "4.3.2",
-    "@angular/platform-browser": "4.3.2",
-    "@angular/platform-browser-dynamic": "4.3.2",
-    "@angular/router": "4.0.0",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-alpha.23",
-    "angular2-qrcode": "2.0.1",
-    "bootstrap": "4.0.0-alpha.6",
+    "@angular/animations": "^13.1.1",
+    "@angular/common": "^13.1.1",
+    "@angular/compiler": "^13.1.1",
+    "@angular/core": "^13.1.1",
+    "@angular/forms": "^13.1.1",
+    "@angular/localize": "^13.1.1",
+    "@angular/platform-browser": "^13.1.1",
+    "@angular/platform-browser-dynamic": "^13.1.1",
+    "@angular/platform-server": "^13.1.1",
+    "@angular/router": "^13.1.1",
+    "@ng-bootstrap/ng-bootstrap": "^11.0.0",
+    "bootstrap": "4.6.1",
     "core-js": "2.4.1",
-    "dragula": "3.7.2",
     "event-source-polyfill": "0.0.7",
     "font-awesome": "4.7.0",
     "luxon": "^1.4.4",
-    "ng2-dragula": "1.3.1",
-    "node-sass": "^4.13.1",
-    "rxjs": "5.1.0",
+    "rxjs": "^6.6.7",
     "ts-helpers": "1.1.1",
+    "tslib": "^1.9.0",
     "virtual-keyboard": "1.26.26",
-    "zone.js": "0.8.4"
+    "zone.js": "0.11.4",
+    "angular2-qrcode": "2.0.1"
   },
   "devDependencies": {
-    "@angular/cli": "1.2.6",
-    "@angular/compiler-cli": "^4.0.0",
+    "@angular-devkit/build-angular": "~13.1.2",
+    "@angular/cli": "^13.1.2",
+    "@angular/compiler-cli": "^13.1.1",
     "@types/jasmine": "2.5.38",
     "@types/jquery": "<2.2",
-    "@types/node": "~6.0.60",
+    "@types/node": "^16.11.13",
     "@types/virtual-keyboard": "1.26.0",
-    "codelyzer": "~2.0.0",
     "extract-text-webpack-plugin": "^3.0.2",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-remap-istanbul": "^0.2.1",
-    "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "tslint": "~6.1.3",
+    "typescript": "^4.5.4"
   }
 }
