@@ -1,23 +1,15 @@
 import {BrowserModule} from "@angular/platform-browser";
 import {NgModule} from "@angular/core";
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import {HttpModule} from "@angular/http";
+import {HttpClientModule} from "@angular/common/http";
 import {RouterModule} from "@angular/router";
 import {AppComponent} from "./app.component";
-import {UserEditComponent} from "./components/user/edit/user-edit.component";
-import {UserViewComponent} from "./components/user/view/user-view.component";
-import {UserService} from "./components/user/user.service";
-import {UserListComponent} from "./components/user/list/user-list.component";
 import {EventListComponent} from "./components/event/list/event-list.component";
 import {EventService} from "./shared/event/event.service";
-import {EventConfigurationComponent} from "./components/event/configuration/event-configuration.component";
-import {PrinterListComponent} from "./components/printer/list/printer-list.component";
 import {PrinterService} from "./components/printer/printer.service";
 import {CloseDetailComponent} from "./components/close-detail/close-detail.component";
 import {KeysPipe} from "./keys.pipe";
 import {WindowRef} from "./window.service";
-import {DragulaModule} from "ng2-dragula";
-import {UserNotifierService} from "./components/user/user-notifier.service";
 import { LoadingIndicatorComponent } from './components/loading-indicator/loading-indicator.component';
 import { ScanLogEntriesComponent } from './components/scan-log-entries/scan-log-entries.component';
 import {ScanLogService} from "./components/scan-log-entries/scan-log.service";
@@ -27,7 +19,6 @@ import {FilterScanLogEntries} from "./filter.pipe";
 import {NgbModule} from "@ng-bootstrap/ng-bootstrap";
 import {ServerEventsService} from "./server-events.service";
 import { SidebarComponent } from './components/sidebar/sidebar.component';
-import { UserPrinterComponent } from './components/user-printer/user-printer.component';
 import { ResponsiveLayoutComponent } from './components/responsive-layout/responsive-layout.component';
 import { EventListPageComponent } from './components/event/list-page/event-list-page.component';
 import { ConfirmPowerOffComponent } from './components/confirm-power-off/confirm-power-off.component';
@@ -35,22 +26,19 @@ import { CheckInComponent } from './components/check-in/check-in.component';
 import {ScanListenerDirective} from "./components/check-in/scan-listener.directive";
 import {ScanService} from "./scan-module/scan/scan.service";
 import {ScanLogEntryReprintComponent} from "app/components/scan-log-entries/reprint/scan-log-entry-reprint.component";
-import { QRCodeModule } from 'angular2-qrcode';
 import {SettingsComponent} from "./components/settings/settings.component";
 import { VirtualKeyboardDirective } from './virtual-keyboard.directive';
 import {ConfigurationService} from "./shared/configuration/configuration.service";
 import { SystemInfoComponent } from './components/system-info/system-info.component';
 import {SidebarWatchComponent} from "./components/sidebar/sidebar.watch.component";
+import {QRCodeModule} from "angular2-qrcode";
+import {CheckScannerComponent} from "./components/check-scanner/check-scanner.component";
+import {AddBreaksPipe} from "./add-breaks.pipe";
 
 @NgModule({
   declarations: [
     AppComponent,
-    UserEditComponent,
-    UserViewComponent,
-    UserListComponent,
     EventListComponent,
-    EventConfigurationComponent,
-    PrinterListComponent,
     CloseDetailComponent,
     KeysPipe,
     FilterScanLogEntries,
@@ -60,7 +48,6 @@ import {SidebarWatchComponent} from "./components/sidebar/sidebar.watch.componen
     ScanLogComponent,
     ScanLogEntryReprintComponent,
     SidebarComponent,
-    UserPrinterComponent,
     ResponsiveLayoutComponent,
     EventListPageComponent,
     ConfirmPowerOffComponent,
@@ -69,39 +56,35 @@ import {SidebarWatchComponent} from "./components/sidebar/sidebar.watch.componen
     ScanListenerDirective,
     VirtualKeyboardDirective,
     SystemInfoComponent,
-    SidebarWatchComponent
+    SidebarWatchComponent,
+    CheckScannerComponent,
+    AddBreaksPipe
   ],
   imports: [
-    NgbModule.forRoot(),
-    BrowserModule,
-    FormsModule,
-    ReactiveFormsModule,
-    HttpModule,
-    DragulaModule,
-    QRCodeModule,
-    RouterModule.forRoot([
-      { path: '', component: HomeComponent, children: [
-        { path: 'user/new', component: UserEditComponent },
-        { path: 'user/edit/:userId', component: UserEditComponent }
-      ] },
-      { path: 'scan-log', children: [
-        { path: '', redirectTo: 'view', pathMatch: 'full' },
-        { path: 'view', component: ScanLogComponent },
-        { path: 'event/:eventKey/entry/:entryId/reprint', component: ScanLogEntryReprintComponent }
-      ]},
-      { path: 'users-printers', component: UserPrinterComponent, children: [
-        { path: 'user/new', component: UserEditComponent },
-        { path: 'user/edit/:userId', component: UserEditComponent }
-      ]},
-      { path: 'check-in', component: CheckInComponent},
-      { path: 'events', component: EventListPageComponent},
-      { path: 'settings', component: SettingsComponent},
-      { path: 'power-off', component: ConfirmPowerOffComponent},
-      { path: 'system-info', component: SystemInfoComponent }
-
-    ])
+      NgbModule,
+      BrowserModule,
+      FormsModule,
+      ReactiveFormsModule,
+      HttpClientModule,
+      RouterModule.forRoot([
+          {path: '', component: HomeComponent},
+          {
+              path: 'scan-log', children: [
+                  {path: '', redirectTo: 'view', pathMatch: 'full'},
+                  {path: 'view', component: ScanLogComponent},
+                  {path: 'event/:eventKey/entry/:entryId/reprint', component: ScanLogEntryReprintComponent}
+              ]
+          },
+          {path: 'check-in', component: CheckInComponent},
+          {path: 'events', component: EventListPageComponent},
+          {path: 'settings', component: SettingsComponent},
+          {path: 'power-off', component: ConfirmPowerOffComponent},
+          {path: 'system-info', component: SystemInfoComponent},
+          {path: 'check-scanner', component: CheckScannerComponent}
+      ]),
+      QRCodeModule
   ],
-  providers: [UserService, EventService, PrinterService, WindowRef, UserNotifierService, ScanLogService, ServerEventsService, ScanService, ConfigurationService],
+  providers: [EventService, PrinterService, WindowRef, ScanLogService, ServerEventsService, ScanService, ConfigurationService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
