@@ -46,24 +46,25 @@ import javax.imageio.ImageIO
 @RestController
 @RequestMapping("/api/internal/users")
 @Profile("server", "full")
-open class UserApi(private val userRepository: UserRepository,
-                   private val transactionManager: PlatformTransactionManager,
-                   private val passwordGenerator: PasswordGenerator,
-                   private val passwordEncoder: PasswordEncoder,
-                   private val authorityRepository: AuthorityRepository,
-                   @Qualifier("localServerURL") private val localServerUrl: String,
-                   private val gson: Gson,
-                   private val sslKeyExporter: SslKeyExporter) {
+@Deprecated("will be removed very soon")
+class UserApi(private val userRepository: UserRepository,
+              private val transactionManager: PlatformTransactionManager,
+              private val passwordGenerator: PasswordGenerator,
+              private val passwordEncoder: PasswordEncoder,
+              private val authorityRepository: AuthorityRepository,
+              @Qualifier("localServerURL") private val localServerUrl: String,
+              private val gson: Gson,
+              private val sslKeyExporter: SslKeyExporter) {
     private val logger = LoggerFactory.getLogger(UserApi::class.java)
 
     @RequestMapping(value = [""], method = [(RequestMethod.GET)])
-    open fun loadAllOperators(): List<User> = doInTransaction<List<User>>().invoke(transactionManager, {userRepository.findAllOperators()}, {
+    fun loadAllOperators(): List<User> = doInTransaction<List<User>>().invoke(transactionManager, {userRepository.findAllOperators()}, {
         logger.error("error while loading users", it)
         emptyList()
     })
 
     @RequestMapping(value = ["/{userId}"], method = [(RequestMethod.GET)])
-    open fun loadSingleUser(@PathVariable("userId") userId: Int): ResponseEntity<User> = doInTransaction<ResponseEntity<User>>().invoke(transactionManager, {
+    fun loadSingleUser(@PathVariable("userId") userId: Int): ResponseEntity<User> = doInTransaction<ResponseEntity<User>>().invoke(transactionManager, {
         userRepository.findById(userId).map {
             ResponseEntity.ok(it)
         }.orElseGet(fun(): ResponseEntity<User> {
@@ -75,7 +76,7 @@ open class UserApi(private val userRepository: UserRepository,
     })
 
     @RequestMapping(value = ["/"], method = [(RequestMethod.POST)])
-    open fun create(@RequestBody() form: UserForm): ResponseEntity<UserWithPassword> = doInTransaction<ResponseEntity<UserWithPassword>>().invoke(transactionManager, {
+    fun create(@RequestBody() form: UserForm): ResponseEntity<UserWithPassword> = doInTransaction<ResponseEntity<UserWithPassword>>().invoke(transactionManager, {
         val username = form.username
         if(username != null && username.isNotBlank()) {
             ResponseEntity.ok(createNewUser(username).invoke(passwordGenerator, passwordEncoder, userRepository, authorityRepository))
@@ -88,7 +89,7 @@ open class UserApi(private val userRepository: UserRepository,
     })
 
     @RequestMapping(value = ["/{userId}/resetPassword"], method = [(RequestMethod.POST)])
-    open fun resetPassword(@PathVariable("userId") userId: Int): ResponseEntity<UserWithPassword> = doInTransaction<ResponseEntity<UserWithPassword>>().invoke(transactionManager, {
+    fun resetPassword(@PathVariable("userId") userId: Int): ResponseEntity<UserWithPassword> = doInTransaction<ResponseEntity<UserWithPassword>>().invoke(transactionManager, {
         userRepository.findById(userId).map {
             ResponseEntity.ok(updatePassword(it).invoke(passwordGenerator, passwordEncoder, userRepository))
         }.orElseGet(fun(): ResponseEntity<UserWithPassword> {
@@ -101,9 +102,9 @@ open class UserApi(private val userRepository: UserRepository,
 
 
     @RequestMapping(value = ["/{userId}/qr-code"], method = [(RequestMethod.GET)])
-    open fun generateQRCode(@PathVariable("userId") userId: Int,
-                            @RequestParam("password") password: String,//base-64 encoded password
-                            response: HttpServletResponse) {
+    fun generateQRCode(@PathVariable("userId") userId: Int,
+                       @RequestParam("password") password: String,//base-64 encoded password
+                       response: HttpServletResponse) {
         doInTransaction<Unit>().invoke(transactionManager, {
             val user = userRepository.findById(userId)
             if(user.isPresent) {
