@@ -16,13 +16,13 @@
  */
 
 
-import {Component, Input, OnInit} from "@angular/core";
+import {Component, OnInit} from "@angular/core";
 import {ProgressManager} from "../../../ProgressManager";
 import {Printer, PrinterService} from "../../printer/printer.service";
-import {ConfigurableLabelContent, ScanLogEntry, ScanLogService} from "../scan-log.service";
+import {ConfigurableLabelContent, ScanLogService} from "../scan-log.service";
 import {EventService, Event} from "../../../shared/event/event.service";
-import {Observable} from "rxjs/Observable";
 import {ActivatedRoute, Params} from "@angular/router";
+import { forkJoin } from "rxjs";
 
 @Component({
   selector: 'scan-log-entry-reprint',
@@ -56,10 +56,10 @@ export class ScanLogEntryReprintComponent implements OnInit {
   private loadData() {
     this.progressManager
       .monitorCall(() => {
-        return Observable.forkJoin(this.scanLogService.getReprintPreview(this.entryId, this.eventKey),
+        return forkJoin([this.scanLogService.getReprintPreview(this.entryId, this.eventKey),
           this.printerService.loadAllPrinters(),
           this.eventService.getSingleEvent(this.eventKey)
-        );
+        ]);
       })
       .subscribe(res => {
         let [content, printers, event] = res;
