@@ -1,9 +1,10 @@
+
+import {map} from 'rxjs/operators';
 import {Component, Input, OnInit, OnDestroy} from "@angular/core";
 import {ScanLogEntry, ScanLogService} from "./scan-log.service";
 import {ProgressManager} from "../../ProgressManager";
-import {Observable, Subscription} from "rxjs";
+import {forkJoin, Subscription} from "rxjs";
 import {Event, EventService} from "../../shared/event/event.service";
-import "rxjs/add/operator/map";
 import {Printer, PrinterService} from "../printer/printer.service";
 import {EventType, ServerEventsService} from "../../server-events.service";
 
@@ -56,12 +57,12 @@ export class ScanLogEntriesComponent implements OnInit, OnDestroy {
   private loadData() {
     this.progressManager
       .monitorCall(() => {
-        return Observable.forkJoin(this.scanLogService.getEntries(this.currentPage - 1, this.pageSize, this.term),
+        return forkJoin([this.scanLogService.getEntries(this.currentPage - 1, this.pageSize, this.term),
           this.eventService.getAllEvents(),
           this.printerService.loadAllPrinters()
-        );
-      })
-      .map(res => {
+        ]);
+      }).pipe(
+      map(res => {
         let [entries, events, printers] = res;
         this.printers = printers.filter(p => p.active);
         this.found = entries.found;
@@ -69,7 +70,7 @@ export class ScanLogEntriesComponent implements OnInit, OnDestroy {
           const entry = entryWithBoxClass.scanLog;
           return new ScanLogEntryWithEvent(entry, events.find(e => e.key === entry.eventKey), entryWithBoxClass.boxColorClass);
         })
-      })
+      }))
       .subscribe(entries => {
         this.entries = entries
       });
